requirements.txt
braintree==3.25.0
Django==1.8.4
django-cors-headers==1.1.0
django-crispy-forms==1.6.0
django-filter==0.13.0
django-registration-redux==1.3
djangorestframework==3.4.3
djangorestframework-jwt==1.8.0
gunicorn==19.6.0
Markdown==2.6.6
Pillow==3.1.1
psycopg2==2.6.1
PyJWT==1.4.1
requests==2.9.1
